ipq807x: board: Pass UUID of to kernel in bootargs in-order to mount rootfs
This change is for mounting the rootfs partition in kernel based
on UUID(Universally Unique IDentifier). In embedded system it is
common to store rootfs on SD card or eMMC. Typically the kernel
command line has like: root=/dev/mmcblk0p20 to tell the kernel
where to look for rootfs partition. The problem in this approach
is the eMMC device number can vary based on wether the SD card
present or not in the SD card slot. Also depending on kernel version
the eMMC device number can vary. This means passing the eMMC device
for specifying the rootfs location is not robust approach.
If SD card first get detected then the base minor 0 is assigned to
SD card and kernel will try to mount rootfs from SD card and if
there is no rootfs present on SD card then kernel fail to mount the
rootfs and we will get the error.

error:
[ 4.562519] VFS: Cannot open root device "(null)" or
unknown-block(179,20): error -6
[ 4.562541] Please append a correct "root=" boot option;

Change-Id: Ifc6ae031facde1a3ac6822dab4bfda1745e44bd0
Signed-off-by: Md Sadre Alam <mdalam@codeaurora.org>